+++ main.tf
@@ -4,6 +4,9 @@ locals {
     local.subdirectory_index_association,
     local.hsts_header_association,
   )
+  content_bucket_name = coalesce(var.content_bucket_name, "${var.domain_name}-static-content")
+  content_bucket      = var.create_content_bucket ? aws_s3_bucket.content[0] : data.aws_s3_bucket.content[0]
+
 }
 resource "aws_route53_record" "record" {
   count   = var.hosted_zone_id != "" ? 1 : 0
@@ -38,27 +41,32 @@ resource "aws_s3_bucket" "logging" {
     Name = "CloudFront logs for ${var.domain_name}"
   }
 }
-resource "aws_s3_bucket" "static" {
-  bucket = "${var.domain_name}-static-content"
+resource "aws_s3_bucket" "content" {
+  count  = var.create_content_bucket ? 1 : 0
+  bucket = local.content_bucket_name
   tags = {
     Name = "${var.domain_name} Static Content"
   }
 }
+data "aws_s3_bucket" "content" {
+  count  = var.create_content_bucket ? 0 : 1
+  bucket = local.content_bucket_name
+}
 # Static Bucket Policy
-data "aws_iam_policy_document" "static" {
+data "aws_iam_policy_document" "content" {
   # Grant read to cloudfront's OAI
   statement {
     actions   = ["s3:GetObject"]
-    resources = ["${aws_s3_bucket.static.arn}/*"]
+    resources = ["${local.content_bucket.arn}/*"]
     principals {
       type        = "AWS"
       identifiers = [aws_cloudfront_origin_access_identity.oai.iam_arn]
     }
   }
 }
-resource "aws_s3_bucket_policy" "static" {
-  bucket = aws_s3_bucket.static.id
-  policy = data.aws_iam_policy_document.static.json
+resource "aws_s3_bucket_policy" "content" {
+  bucket = local.content_bucket.id
+  policy = data.aws_iam_policy_document.content.json
 }
 resource "aws_cloudfront_origin_access_identity" "oai" {
   comment = "OAI for ${var.domain_name}"
@@ -89,8 +97,8 @@ module "cloudfront" {
 
   s3_origins = [
     {
-      origin_id              = "static_bucket"
-      domain_name            = aws_s3_bucket.static.bucket_regional_domain_name
+      origin_id              = local.content_bucket_name
+      domain_name            = local.content_bucket.bucket_regional_domain_name
       origin_access_identity = aws_cloudfront_origin_access_identity.oai.cloudfront_access_identity_path
     },
   ]
@@ -99,7 +107,7 @@ module "cloudfront" {
   default_cache_behavior = {
     allowed_methods                = ["GET", "HEAD", "OPTIONS", "PUT", "POST", "PATCH", "DELETE"]
     cached_methods                 = ["GET", "HEAD"]
-    origin_id                      = "static_bucket"
+    origin_id                      = local.content_bucket_name
     default_ttl                    = 0
     min_ttl                        = 0
     max_ttl                        = 0