+++ outputs.tf
@@ -1,4 +1,4 @@
 output "content_bucket_id" {
-  value       = aws_s3_bucket.static.id
+  value       = local.content_bucket.id
   description = "The bucket where you will put your static html content"
 }